pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "physics-ai-research-tool"
version = "0.1.0"
description = "Autonomous theoretical physics research engine for modified gravity black-hole studies."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Physics AI Research Tool" }
]
dependencies = [
  "pydantic>=2.8,<3",
  "sympy>=1.12",
  "numpy>=1.26",
  "scipy>=1.12",
  "typer>=0.12",
  "pyyaml>=6.0",
  "pandas>=2.2",
  "matplotlib>=3.9",
  "sqlalchemy>=2.0",
  "httpx>=0.27",
]

[project.optional-dependencies]
hpc = [
  "dask[distributed]>=2024.5.0",
]
vector = [
  "faiss-cpu>=1.8.0",
]
llm = [
  "openai>=1.40.0",
]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.6.0",
  "mypy>=1.11",
]

[project.scripts]
physai = "physics_ai.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_configs = true
